Refactor Connection, HostPool, and Strategy to ease mocking
FakeStrategy is now also removed, since it would need a lot of configuration and it is easier to just mock the actual Strategy now that it is neat and refactored.

We still need to make an easy way to mock out the actual LDAP results instead of just the warden-ldap results.